Move eventHasSameFields and createStripped entities methods to ImportExportUtils.ts
