+++ src/calendar-app/calendar/model/CalendarModel.ts
@@ -20,7 +20,6 @@ import { EventController } from "../../../common/api/main/EventController"
 import {
 	createDateWrapper,
 	createMembershipRemoveData,
-	DateWrapper,
 	Group,
 	GroupInfo,
 	GroupInfoTypeRef,
@@ -31,18 +30,14 @@ import {
 	UserAlarmInfoTypeRef,
 } from "../../../common/api/entities/sys/TypeRefs.js"
 import {
-	AdvancedRepeatRule,
 	CalendarEvent,
-	CalendarEventAttendee,
 	CalendarEventTypeRef,
 	CalendarEventUpdate,
 	CalendarEventUpdateTypeRef,
 	CalendarGroupRoot,
 	CalendarGroupRootTypeRef,
-	CalendarRepeatRule,
 	createDefaultAlarmInfo,
 	createGroupSettings,
-	EncryptedMailAddress,
 	FileTypeRef,
 	GroupSettings,
 	UserSettingsGroupRootTypeRef,
@@ -65,7 +60,6 @@ import {
 	listIdPart,
 	POST_MULTIPLE_LIMIT,
 	removeTechnicalFields,
-	Stripped,
 } from "../../../common/api/common/utils/EntityUtils"
 import type { AlarmScheduler } from "../../../common/calendar/date/AlarmScheduler.js"
 import { Notifications, NotificationType } from "../../../common/gui/Notifications"
@@ -104,6 +98,7 @@ import { ExternalCalendarFacade } from "../../../common/native/common/generatedi
 import { deviceConfig, DeviceConfig } from "../../../common/misc/DeviceConfig.js"
 import { locator } from "../../../common/api/main/CommonLocator.js"
 import {
+	eventHasSameFields,
 	EventImportRejectionReason,
 	EventWrapper,
 	parseCalendarStringData,
@@ -154,63 +149,6 @@ export function assertEventValidity(event: CalendarEvent) {
 			throw new UserError("pre1970Start_msg")
 		case CalendarEventValidity.Valid:
 		// event is valid, nothing to do
-	}
-}
-
-type StrippedRepeatRule = Stripped<
-	Omit<CalendarRepeatRule, "excludedDates" | "advancedRules"> & {
-		excludedDates: Stripped<DateWrapper>[]
-		advancedRules: Stripped<AdvancedRepeatRule>[]
-	}
->
-
-function createStrippedRepeatRule(repeatRule: CalendarRepeatRule | null): StrippedRepeatRule | null {
-	if (!repeatRule) {
-		return null
-	}
-	return {
-		frequency: repeatRule.frequency ?? "",
-		endType: repeatRule.endType ?? "",
-		endValue: repeatRule.endValue ?? "",
-		interval: repeatRule.interval ?? "",
-		timeZone: repeatRule.timeZone ?? "",
-		excludedDates: repeatRule.excludedDates
-			? repeatRule.excludedDates.map((ex) => ({
-					date: ex.date,
-			  }))
-			: [],
-		advancedRules: repeatRule.advancedRules
-			? repeatRule.advancedRules.map((rule) => ({
-					ruleType: rule.ruleType,
-					interval: rule.interval,
-			  }))
-			: [],
-	}
-}
-
-type StrippedCalendarEventAttendee = Stripped<
-	Omit<CalendarEventAttendee, "address"> & {
-		address: Stripped<EncryptedMailAddress>
-	}
->
-
-function createStrippedAttendees(attendees: CalendarEventAttendee[]): StrippedCalendarEventAttendee[] {
-	return attendees.map((attendee: CalendarEventAttendee) => {
-		return {
-			status: attendee.status,
-			address: createStrippedMailAddress(attendee.address)!,
-		}
-	})
-}
-
-function createStrippedMailAddress(mailAddress: EncryptedMailAddress | null): Stripped<EncryptedMailAddress> | null {
-	if (!mailAddress) {
-		return null
-	}
-
-	return {
-		address: mailAddress.address,
-		name: mailAddress.name,
 	}
 }
 
@@ -530,7 +468,7 @@ export class CalendarModel {
 					return false
 				}
 
-				return !this.eventHasSameFields(event, existingEvent)
+				return !eventHasSameFields(event, existingEvent)
 			})
 
 			const eventsToRemove = existingEventList.filter(
@@ -642,7 +580,7 @@ export class CalendarModel {
 				continue
 			}
 
-			if (this.eventHasSameFields(duplicatedEvent, existingEvent)) {
+			if (eventHasSameFields(duplicatedEvent, existingEvent)) {
 				continue
 			}
 			await this.updateEventWithExternal(existingEvent, duplicatedEvent)
@@ -671,29 +609,6 @@ export class CalendarModel {
 			await this.calendarFacade.saveImportedCalendarEvents(eventsForCreation, 0)
 		}
 		console.log(TAG, `${operationsLog.created} events created`)
-	}
-
-	// visible for testing
-	public eventHasSameFields(a: CalendarEvent, b: CalendarEvent) {
-		const rruleA = createStrippedRepeatRule(a.repeatRule)
-		const rruleB = createStrippedRepeatRule(b.repeatRule)
-		const attendeesA = createStrippedAttendees(a.attendees)
-		const attendeesB = createStrippedAttendees(b.attendees)
-		const organizerA = createStrippedMailAddress(a.organizer)
-		const organizerB = createStrippedMailAddress(b.organizer)
-
-		return (
-			a.startTime.valueOf() === b.startTime.valueOf() &&
-			a.endTime.valueOf() === b.endTime.valueOf() &&
-			deepEqual({ ...attendeesA }, { ...attendeesB }) &&
-			a.summary === b.summary &&
-			a.sequence === b.sequence &&
-			a.location === b.location &&
-			a.description === b.description &&
-			deepEqual(organizerA, organizerB) &&
-			deepEqual(rruleA, rruleB) &&
-			a.recurrenceId?.valueOf() === b.recurrenceId?.valueOf()
-		)
 	}
 
 	private async loadOrCreateCalendarInfo(progressMonitor: IProgressMonitor): Promise<ReadonlyMap<Id, CalendarInfo>> {
